perf(decode): give the mixed co-dispatch head the same lm_head ladder
`decode_b2::mixed_final_norm_lm_head` — the prefill+decode co-dispatch head
reached from `decode_b.rs` via `mixed_forward_dispatch`, i.e. LIVE
continuous-batching traffic — still looped `padded_n` times through
`ops::w4a16_gemv`, streaming the entire vocab weight once PER ROW. On a
[248320, 5120] NVFP4 head that is ~N x 254 MB/step. Its sibling `decode_a2`
grew the batched ladder weeks ago; the two had simply drifted.

SITE CREDIT: @rsafier, #332. That PR found exactly the right line. It fixed it
with a bare `batch16` behind a default-OFF env var; this takes the other
approach.

WHY EXTRACT RATHER THAN COPY. Duplicating the ladder would put the choice of
"which kernel does padded_n land on" in two independently-maintained places.
The first thing to go wrong would not be a perf drift — it would be a silent
NUMERICAL difference between the pure-decode and co-dispatch paths at the same
batch width, which is far worse than the bandwidth bug being fixed. So
`decode_a2`'s block moves to `trait_impl/lm_head_batched.rs` as one
`lm_head_project_batched()`, and both heads call it. `v` is read from
`self.config.vocab_size` inside rather than passed, so there is no second place
for it to be wrong.

Default-ON with the existing kill-switch (`ATLAS_NO_LM_HEAD_BATCH_GEMV=1`),
inherited unchanged — not a new flag. Per house rule: ship wins as defaults
with a kill-switch, not as opt-in.

Side effect worth having: `decode_a2.rs` drops 617 -> 501 LoC.

spark-model: 311 passed, 0 failed. fmt + clippy clean, no LoC violations.

★ HONESTY ABOUT THE CLAIM. This is a BANDWIDTH-ACCOUNTING argument, not a
measurement. Nobody has A/B'd the mixed path, and I am not claiming a number.
The gate run this PR owes must be spec-OFF or on reversed-order pairs — a
single spec-ON pair drifts +/-2%, which is wider than any effect worth
reporting here.

NOT INCLUDED, deliberately: `SequenceState::cached_prefix_blocks`
(`traits.rs:140`) is written at three sites and read at none. Removing a `pub`
field is an API change and does not belong bundled into a perf PR; noted for a
separate one.
