+++ crates/spark-model/src/model/trait_impl/decode_b2.rs
@@ -55,46 +55,18 @@ impl TransformerModel {
             stream,
         )?;
 
-        let logits = self.buffers.logits();
+        // The SAME ladder the pure-decode head uses (`lm_head_batched.rs`).
+        //
+        // This loop used to run `padded_n` separate `w4a16_gemv` calls, each
+        // streaming the entire vocab weight — ~N x 254 MB/step on a
+        // [248320, 5120] NVFP4 head, on live continuous-batching traffic,
+        // while `decode_a2` had had the batched ladder for weeks. Sharing one
+        // function is what keeps the two heads from picking different kernels
+        // (and therefore different numerics) at the same `padded_n`.
+        //
+        // Site credit: @rsafier, #332.
+        let logits = self.lm_head_project_batched(normed, padded_n, h, bf16, stream)?;
         let v = self.config.vocab_size;
-        for i in 0..padded_n {
-            let normed_i = normed.offset(i * h * bf16);
-            let logits_i = logits.offset(i * v * bf16);
-            if let Some(ref fp8) = self.lm_head_fp8 {
-                ops::dense_gemv_fp8w(
-                    self.gpu.as_ref(),
-                    self.dense_gemv_fp8w_kernel,
-                    normed_i,
-                    fp8,
-                    logits_i,
-                    v as u32,
-                    h as u32,
-                    stream,
-                )?;
-            } else if let Some(ref nvfp4) = self.lm_head_nvfp4 {
-                ops::w4a16_gemv(
-                    self.gpu.as_ref(),
-                    self.w4a16_gemv_kernel,
-                    normed_i,
-                    nvfp4,
-                    logits_i,
-                    v as u32,
-                    h as u32,
-                    stream,
-                )?;
-            } else {
-                ops::dense_gemv(
-                    self.gpu.as_ref(),
-                    self.dense_gemv_kernel,
-                    normed_i,
-                    &self.lm_head_weight,
-                    logits_i,
-                    v as u32,
-                    h as u32,
-                    stream,
-                )?;
-            }
-        }
         let decode_logits = logits;
 
         // 7b. Prefill logits: norm last token → 1 GEMV (if is_last)
