+++ crates/spark-model/src/model/trait_impl/mod.rs
@@ -26,6 +26,7 @@ mod decode_checkpoint;
 mod decode_graph_key;
 mod drafter_prefill;
 mod ep_misc;
+mod lm_head_batched;
 mod meta;
 mod prefill_a;
 mod prefill_b;
